move test transformer to test cases
